Fixed UI bugs. Renamed progress-range to time-range.
Updated examples.
