+++ examples/control-elements/media-controller.html
@@ -1,7 +1,7 @@
 <html>
 <head>
   <meta name="viewport" content="width=device-width">
-  <script type="module" src="../dist/index.js"></script>
+  <script type="module" src="../../dist/index.js"></script>
 </head>
 <body>
