+++ examples/control-elements/media-time-range.html
@@ -1,0 +1,23 @@
+<html>
+<head>
+  <meta name="viewport" content="width=device-width">
+  <script type="module" src="../../dist/index.js"></script>
+</head>
+<body>
+
+<h1>&lt;media-time-range&gt;</h1>
+
+<h3>Default (no time or duration)</h3>
+<media-time-range></media-time-range>
+
+<h3>Time 0s (duration 60s)</h3>
+<media-time-range media-current-time="0" media-duration="60"></media-time-range>
+
+<h3>Time 30s (duration 60s)</h3>
+<media-time-range media-current-time="30" media-duration="60"></media-time-range>
+
+<h3>Time 60s (duration 60s)</h3>
+<media-time-range media-current-time="60" media-duration="60"></media-time-range>
+
+</body>
+</html>
