+++ examples/control-elements/media-volume-range.html
@@ -1,0 +1,28 @@
+<html>
+<head>
+  <meta name="viewport" content="width=device-width">
+  <script type="module" src="../../dist/index.js"></script>
+</head>
+<body>
+
+<h1>&lt;media-volume-range&gt;</h1>
+
+<h3>Default</h3>
+<media-volume-range></media-volume-range>
+
+<h3>Volume 0</h3>
+<media-volume-range media-volume="0"></media-volume-range>
+
+<h3>Volume 0.5</h3>
+<media-volume-range media-volume="0.5"></media-volume-range>
+
+<h3>Volume 1</h3>
+<media-volume-range media-volume="1"></media-volume-range>
+
+<h3>In control bar</h3>
+<media-control-bar>
+  <media-volume-range></media-volume-range>
+</media-control-bar>
+
+</body>
+</html>
