+++ examples/index.html
@@ -28,8 +28,13 @@
 
 <h2>Control element examples</h2>
 <ul>
-  <li><a href="/examples/control-elements/media-controller.html">Clip selector</a></li>
-  <li><a href="/examples/control-elements/clip-selector.html">Clip selector</a></li>
+  <li><a href="/examples/control-elements/media-play-button.html">Play button</a></li>
+  <li><a href="/examples/control-elements/media-mute-button.html">Mute button</a></li>
+  <li><a href="/examples/control-elements/media-volume-range.html">Volume range</a></li>
+  <li><a href="/examples/control-elements/media-clip-selector.html">Clip selector</a></li>
+  <li><a href="/examples/control-elements/media-controller.html">Media controller</a></li>
+  <li><a href="/examples/control-elements/media-time-range.html">Time range</a></li>
+  <li><a href="/examples/control-elements/media-fullscreen-button.html">Fullscreen button</a></li>
 </ul>
 
 </body>
