+++ index.js
@@ -23,6 +23,7 @@ import MediaClipSelector from './src/js/media-clip-selector.js';
 import MediaReplayButton from './src/js/media-replay-button.js';
 import MediaSettingsPopup from './src/js/media-settings-popup.js';
 import MediaThumbnailPreviewElement from './src/js/media-thumbnail-preview-element.js';
+import MediaTimeRange from './src/js/media-time-range.js';
 import MediaTitleElement from './src/js/media-title-element.js';
 import MediaVolumeRange from './src/js/media-volume-range.js';
 import { Window as window } from './src/js/utils/server-safe-globals.js';
@@ -60,6 +61,7 @@ export {
   MediaReplayButton,
   MediaSettingsPopup,
   MediaThumbnailPreviewElement,
+  MediaTimeRange,
   MediaTitleElement,
   MediaVolumeRange,
 }
