+++ src/js/media-chrome-range.js
@@ -42,7 +42,7 @@ template.innerHTML = `
       display: inline-block;
       vertical-align: middle;
       box-sizing: border-box;
-      background-color: transparent;
+      background-color: var(--media-control-background, rgba(20,20,30, 0.7));
       transition: background-color 0.15s linear;
       height: 44px;
       width: 100px;
@@ -61,7 +61,7 @@ template.innerHTML = `
     }
 
     :host(:hover) {
-      background: rgba(255,255,255, 0.10);
+      background-color: var(--media-control-hover-background, rgba(50,50,60, 0.7));
     }
 
     input[type=range] {
